0.4.3: fix debug assertion in mesochronous timer
early iteration caused a debug 't_now.duration_since(t_next)' call to
panic because t_now was before t_next. also made the same change for the
isochronous timer although it was not observed to panic in the same way
it would technically be possible if the thread woke too early from the
last tick.